+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # renv 0.13.0 (UNRELEASED)
 
+* The library used for staged installs can now be configured via the
+  `RENV_PATHS_LIBRARY_STAGING` environment variable. (#584)
+  
 * Fixed an issue where bootstrapping an older version of `renv` could
   fail if the R repositories had not been appropriately set.